updated data type of cusotmersInQueue

+++ backendservice/src/main/java/com/google/gpay/virtualqueue/backendservice/proto/GetShopResponse.java
@@ -22,5 +22,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class GetShopResponse {
     private Shop shop;
-    private Integer customersInQueue;
+    private long customersInQueue;
 }
